requirements.txt
telethon>=1.24.0
python-dotenv>=0.19.0
requests>=2.26.0
requests-toolbelt>=0.9.1